requirements.txt
Django
south
djangorestframework
MySQL-python
xlrd
